# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/teachers": {
      "get": {
        "description": "Paginated; search + status filter. state=deleted|all is owner-only.",
        "operationId": "TeachersController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "exclusiveMinimum": false,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "exclusiveMaximum": false,
              "exclusiveMinimum": false,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "ARCHIVED"
              ]
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "default": "active",
              "type": "string",
              "enum": [
                "active",
                "deleted",
                "all"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherListDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List workspace teachers",
        "tags": [
          "teachers"
        ]
      },
      "post": {
        "operationId": "TeachersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeacherDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a teacher profile",
        "tags": [
          "teachers"
        ]
      }
    },
    "/api/teachers/{teacherId}": {
      "get": {
        "operationId": "TeachersController_getDetail",
        "parameters": [
          {
            "name": "teacherId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a teacher profile",
        "tags": [
          "teachers"
        ]
      },
      "patch": {
        "operationId": "TeachersController_update",
        "parameters": [
          {
            "name": "teacherId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeacherDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a teacher",
        "tags": [
          "teachers"
        ]
      },
      "delete": {
        "description": "Hides the teacher from pickers; enrollments/lessons keep the reference.",
        "operationId": "TeachersController_remove",
        "parameters": [
          {
            "name": "teacherId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft-delete a teacher",
        "tags": [
          "teachers"
        ]
      }
    },
    "/api/teachers/{teacherId}/restore": {
      "post": {
        "operationId": "TeachersController_restore",
        "parameters": [
          {
            "name": "teacherId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restore a soft-deleted teacher",
        "tags": [
          "teachers"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tutorio API",
    "description": "Financial calendar for private tutors",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message"
        ]
      },
...
      "TeacherListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "workspaceId": {
                  "type": "string",
                  "format": "uuid"
                },
                "fullName": {
                  "type": "string"
                },
                "email": {
                  "type": "string",
                  "nullable": true
                },
                "phone": {
                  "type": "string",
                  "nullable": true
                },
                "telegramUsername": {
                  "type": "string",
                  "nullable": true
                },
                "subjects": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "enum": [
                      "MATH",
                      "ENGLISH",
                      "GERMAN",
                      "FRENCH",
                      "POLISH",
                      "UKRAINIAN_LANGUAGE",
                      "UKRAINIAN_LITERATURE",
                      "WORLD_LITERATURE",
                      "PHYSICS",
                      "CHEMISTRY",
                      "BIOLOGY",
                      "GEOGRAPHY",
                      "HISTORY",
                      "WORLD_HISTORY",
                      "HISTORY_OF_UKRAINE",
                      "COMPUTER_SCIENCE",
                      "ECONOMICS",
                      "LAW",
                      "MUSIC",
                      "ART",
                      "PHYSICAL_EDUCATION",
                      "NMT_PREP",
                      "ZNO_PREP",
                      "IELTS_PREP",
                      "TOEFL_PREP",
                      "SAT_PREP"
                    ]
                  }
                },
                "bio": {
                  "type": "string",
                  "nullable": true
                },
                "defaultRateMinor": {
                  "type": "integer",
                  "minimum": 0,
                  "exclusiveMinimum": false,
                  "nullable": true
                },
                "currency": {
                  "type": "string",
                  "enum": [
                    "EUR",
                    "UAH",
                    "PLN",
                    "USD",
                    "GBP"
                  ],
                  "nullable": true
                },
                "color": {
                  "type": "string",
                  "pattern": "^#[0-9a-fA-F]{6}$",
                  "nullable": true
                },
                "avatarKey": {
                  "type": "string",
                  "enum": [
                    "user-1",
                    "user-2",
                    "user-3",
                    "user-4",
                    "user-5",
                    "user-6",
                    "user-7",
                    "user-8",
                    "user-9",
                    "user-10"
                  ],
                  "nullable": true
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "ACTIVE",
                    "ARCHIVED"
                  ]
                },
                "workspaceMemberId": {
                  "type": "string",
                  "format": "uuid",
                  "nullable": true
                },
                "notes": {
                  "type": "string",
                  "nullable": true
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "deletedAt": {
                  "type": "string",
                  "format": "date-time",
                  "nullable": true
                },
                "activeEnrollmentCount": {
                  "type": "integer",
                  "minimum": 0,
                  "exclusiveMinimum": false
                }
              },
              "required": [
                "id",
                "workspaceId",
                "fullName",
                "email",
                "phone",
                "telegramUsername",
                "subjects",
                "bio",
                "defaultRateMinor",
                "currency",
                "color",
                "avatarKey",
                "status",
                "workspaceMemberId",
                "notes",
                "createdAt",
                "updatedAt",
                "deletedAt",
                "activeEnrollmentCount"
              ]
            }
          },
          "page": {
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false
          },
          "total": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
      "CreateTeacherDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "minLength": 3,
            "maxLength": 32,
            "pattern": "^\\+?[\\d ()-]+$"
          },
          "telegramUsername": {
            "type": "string",
            "maxLength": 64
          },
          "subjects": {
            "type": "array",
            "maxItems": 30,
            "items": {
              "type": "string",
              "enum": [
                "MATH",
                "ENGLISH",
                "GERMAN",
                "FRENCH",
                "POLISH",
                "UKRAINIAN_LANGUAGE",
                "UKRAINIAN_LITERATURE",
                "WORLD_LITERATURE",
                "PHYSICS",
                "CHEMISTRY",
                "BIOLOGY",
                "GEOGRAPHY",
                "HISTORY",
                "WORLD_HISTORY",
                "HISTORY_OF_UKRAINE",
                "COMPUTER_SCIENCE",
                "ECONOMICS",
                "LAW",
                "MUSIC",
                "ART",
                "PHYSICAL_EDUCATION",
                "NMT_PREP",
                "ZNO_PREP",
                "IELTS_PREP",
                "TOEFL_PREP",
                "SAT_PREP"
              ]
            }
          },
          "bio": {
            "type": "string",
            "maxLength": 2000
          },
          "defaultRateMinor": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false,
            "maximum": 2147483647,
            "exclusiveMaximum": false
          },
          "currency": {
            "type": "string",
            "enum": [
              "EUR",
              "UAH",
              "PLN",
              "USD",
              "GBP"
            ]
          },
          "color": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$"
          },
          "avatarKey": {
            "type": "string",
            "enum": [
              "user-1",
              "user-2",
              "user-3",
              "user-4",
              "user-5",
              "user-6",
              "user-7",
              "user-8",
              "user-9",
              "user-10"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "ARCHIVED"
            ],
            "default": "ACTIVE"
          },
          "workspaceMemberId": {
            "type": "string",
            "format": "uuid"
          },
          "notes": {
            "type": "string",
            "maxLength": 2000
          }
        },
        "required": [
          "fullName"
        ]
      },
      "TeacherDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "telegramUsername": {
            "type": "string",
            "nullable": true
          },
          "subjects": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "MATH",
                "ENGLISH",
                "GERMAN",
                "FRENCH",
                "POLISH",
                "UKRAINIAN_LANGUAGE",
                "UKRAINIAN_LITERATURE",
                "WORLD_LITERATURE",
                "PHYSICS",
                "CHEMISTRY",
                "BIOLOGY",
                "GEOGRAPHY",
                "HISTORY",
                "WORLD_HISTORY",
                "HISTORY_OF_UKRAINE",
                "COMPUTER_SCIENCE",
                "ECONOMICS",
                "LAW",
                "MUSIC",
                "ART",
                "PHYSICAL_EDUCATION",
                "NMT_PREP",
                "ZNO_PREP",
                "IELTS_PREP",
                "TOEFL_PREP",
                "SAT_PREP"
              ]
            }
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "defaultRateMinor": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false,
            "nullable": true
          },
          "currency": {
            "type": "string",
            "enum": [
              "EUR",
              "UAH",
              "PLN",
              "USD",
              "GBP"
            ],
            "nullable": true
          },
          "color": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "nullable": true
          },
          "avatarKey": {
            "type": "string",
            "enum": [
              "user-1",
              "user-2",
              "user-3",
              "user-4",
              "user-5",
              "user-6",
              "user-7",
              "user-8",
              "user-9",
              "user-10"
            ],
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "ARCHIVED"
            ]
          },
          "workspaceMemberId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "workspaceId",
          "fullName",
          "email",
          "phone",
          "telegramUsername",
          "subjects",
          "bio",
          "defaultRateMinor",
          "currency",
          "color",
          "avatarKey",
          "status",
          "workspaceMemberId",
          "notes",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "UpdateTeacherDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "minLength": 3,
            "maxLength": 32,
            "pattern": "^\\+?[\\d ()-]+$",
            "nullable": true
          },
          "telegramUsername": {
            "type": "string",
            "maxLength": 64,
            "nullable": true
          },
          "subjects": {
            "type": "array",
            "maxItems": 30,
            "items": {
              "type": "string",
              "enum": [
                "MATH",
                "ENGLISH",
                "GERMAN",
                "FRENCH",
                "POLISH",
                "UKRAINIAN_LANGUAGE",
                "UKRAINIAN_LITERATURE",
                "WORLD_LITERATURE",
                "PHYSICS",
                "CHEMISTRY",
                "BIOLOGY",
                "GEOGRAPHY",
                "HISTORY",
                "WORLD_HISTORY",
                "HISTORY_OF_UKRAINE",
                "COMPUTER_SCIENCE",
                "ECONOMICS",
                "LAW",
                "MUSIC",
                "ART",
                "PHYSICAL_EDUCATION",
                "NMT_PREP",
                "ZNO_PREP",
                "IELTS_PREP",
                "TOEFL_PREP",
                "SAT_PREP"
              ]
            }
          },
          "bio": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true
          },
          "defaultRateMinor": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false,
            "maximum": 2147483647,
            "exclusiveMaximum": false,
            "nullable": true
          },
          "currency": {
            "type": "string",
            "enum": [
              "EUR",
              "UAH",
              "PLN",
              "USD",
              "GBP"
            ],
            "nullable": true
          },
          "color": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "nullable": true
          },
          "avatarKey": {
            "type": "string",
            "enum": [
              "user-1",
              "user-2",
              "user-3",
              "user-4",
              "user-5",
              "user-6",
              "user-7",
              "user-8",
              "user-9",
              "user-10"
            ],
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "ARCHIVED"
            ]
          },
          "workspaceMemberId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true
          }
        }
      },
...
    }
  }
}